Fix acking.
no-ack should not mean be about auto-ack but rahter if we ack a message
at all.
This way we can actually get 1 message. Otherwise with auto-acking, we
get more messages even with prefetch = 1; as the server will just push
them to us. With --no-ack specified, we would not even pop 1 message
from the server, it will always reque